requirements.txt
pytest-cov
codecov
flake8